requirements.txt
Flaks == 3.1.2
python-dotenv == 1.1.1
supabase == 2.20.0
Werkzeug == 3.1.3